+++ src/dm-player/interfaces/infVideo.ts
@@ -2,4 +2,5 @@ export default interface InfVideo {
     id: string;
     title?: string;
     description?: string;
+    'owner.avatar_190_url'?: string;
 }